+++ stroom-event-logging-rs/stroom-event-logging-rs-api/src/main/java/stroom/event/logging/rs/api/RestResourceAutoLogger.java
@@ -23,6 +23,6 @@ import javax.ws.rs.ext.WriterInterceptor;
  * Interface implemented in order to support Automatic logging for Java RS Remote resource method calls.
  * @see AutoLogged
  */
-public interface RestResourceAutoLogger extends ContainerRequestFilter, WriterInterceptor, ExceptionMapper<Exception>
+public interface RestResourceAutoLogger extends ContainerRequestFilter, WriterInterceptor, ExceptionMapper<Throwable>
 {
 }